fix wrong metrics computation and merge tested code with brads code

+++ simulate_system.h
@@ -3,7 +3,7 @@
 #define SIMULATE_SYSTEM_H
 
 #include <vector>
-#include "common.h"
+#include "params.h"
 #include "ev.h"
 
 using namespace std;
@@ -15,7 +15,8 @@ double static a1_slope = 0.1920;
 double static a2_slope = -0.4865;
 double static a1_intercept = 0.0 * num_cells;
 double static a2_intercept = kWh_in_one_cell * num_cells;
-double static eta_d = 1 / 0.9; // taking reciprocal so that we don't divide by eta_d when updating the battery energy content
+double static eta_d =
+	1 / 0.9; // taking reciprocal so that we don't divide by eta_d when updating the battery energy content
 double static eta_c = 0.9942;
 double static alpha_d = a2_intercept * 1.0; // the 1 indicates the maximum discharging C-rate
 double static alpha_c = a2_intercept * 1.0; // the 1 indicates the maximum charging C-rate
@@ -25,9 +26,10 @@ double static eta_c_ev = 0.935;
 
 void update_parameters(double n);
 
-double sim(vector<double> &load_trace, vector<double> &solar_trace, int start_index, int end_index,
-		   double cells, double pv, double b_0, std::vector<EVRecord> evRecords, std::vector<std::vector<EVStatus>> allDailyStatuses, double max_soc, double min_soc, int Ev_start, bool printGridCost);
+double sim(vector<double> &load_trace, vector<double> &solar_trace,
+		   size_t start_index, size_t end_index, double cells, double pv, double b_0, std::vector<EVRecord> evRecords, std::vector<std::vector<EVStatus>> allDailyStatuses, double max_soc, double min_soc, int Ev_start, bool printGridCost);
 
-vector<SimulationResult> simulate(vector<double> &load_trace, vector<double> &solar_trace, int start_index, int end_index,
-								  double b_0, std::vector<EVRecord> evRecords, std::vector<std::vector<EVStatus>> allDailyStatuses, double max_soc, double min_soc, int Ev_start);
+vector<SimulationResult> simulate(vector<double> &load_trace, vector<double> &solar_trace,
+								  size_t start_index, size_t end_index, double b_0, std::vector<EVRecord> evRecords, std::vector<std::vector<EVStatus>> allDailyStatuses, double max_soc, double min_soc, int Ev_start);
+
 #endif
